refactor: 자판기 투입 금액, 이벤트 알림 상태 APP으로 위치 변경
지갑 컴포넌트에서도 자판기에 투입되는 액션이 필요하기 때문에 기존에 자판기에서 선언된 상태를 APP으로 위치 변경
